requirements.txt
requests==2.28.1
python-dotenv==0.21.0
pillow==9.2.0
python-dateutil==2.8.2
